backport of commit 72b7aff57570204ca9e04927ac23831a61a69ab9
