+++ nomad/structs/alloc_test.go
@@ -13,6 +13,18 @@ import (
 	"github.com/shoenig/test/must"
 )
 
+// TestAllocation_Canonicalize_New asserts that an alloc with latest
+// schema isn't modified with Canonicalize
+func TestAllocation_Canonicalize_New(t *testing.T) {
+	ci.Parallel(t)
+
+	alloc := MockAlloc()
+	copy := alloc.Copy()
+
+	alloc.Canonicalize()
+	must.Eq(t, copy, alloc)
+}
+
 func TestAllocServiceRegistrationsRequest_StaleReadSupport(t *testing.T) {
 	req := &AllocServiceRegistrationsRequest{}
 	must.True(t, req.IsRead())
@@ -633,5 +645,26 @@ func TestAllocation_NextRescheduleTime(t *testing.T) {
 				must.False(t, eligible)
 			}
 		})
+	}
+}
+
+func TestAllocation_Index(t *testing.T) {
+	ci.Parallel(t)
+
+	a1 := Allocation{
+		Name:      "example.cache[1]",
+		TaskGroup: "cache",
+		JobID:     "example",
+		Job: &Job{
+			ID:         "example",
+			TaskGroups: []*TaskGroup{{Name: "cache"}}},
+	}
+	e1 := uint(1)
+	a2 := a1.Copy()
+	a2.Name = "example.cache[713127]"
+	e2 := uint(713127)
+
+	if a1.Index() != e1 || a2.Index() != e2 {
+		t.Fatalf("Got %d and %d", a1.Index(), a2.Index())
 	}
 }
